+++ dbt.js
@@ -1,5 +1,6 @@
 const Database = require('better-sqlite3')
 const { CommandoClient } = require(`discord.js-commando`)
+const token = require(`${process.cwd()}/config/token.json`)
 
 const path = require(`path`)
 
@@ -26,12 +27,13 @@ client.registry
     // [`pc`, `General stuff about computers.`],
     // [`economy`, `Economy related commands.`],
     // [`animals`, `All sorts of animal related commands.`],
-    [`fun`, `Very fun commands ;)`],
+    //[`fun`, `Very fun commands ;)`],
     /* [`mods`, `Moderation related commands.`], */
     // [`owner`, `Owner-only commands.`],
-    [`info`, `Informative commands.`],
-    [`userdata`, `Shows data about a particular user.`],
-    [`moderation`, `An easy way to moderate a server.`]
+    //[`info`, `Informative commands.`],
+    //[`userdata`, `Shows data about a particular user.`],
+    [`moderation`, `Moderation related commands.`],
+    [`bot`, `Commands that allow you to manipulate the bot without the command line.`]
   ])
   .registerDefaultGroups()
   .registerDefaultCommands({
@@ -69,5 +71,5 @@ fs.readdir(`./events/`, (err, files) => {
   })
 })
 
-client.login(process.env.BOT_TOKEN);
+client.login(token.token);