+++ scripts/start.sh
@@ -1,0 +1,2 @@
+yarn
+pm2 start config/pm2.config.js