+++ changelog.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), adapted to my personal feelings because theres a lot of stuff that I am too lazy to document.
 Most recent changes are located at the top.
 
+## [1.7.0]
+
+### Info:
+    - "The update update"
+
+### Added:
+    - Update command
+    - Restart command
+    - modRoles config functionality
+
+### Fixed:
+    - Needlessly difficult levels
+    
 ## [1.6.0]
 
 ### Info: